allow computer players to be aware of game state
The computer needs knowledge of the game's state
in order to determine how accurate it's guess was.

To promote proper seperation of concerns I wanted
the computer to just read the STDOUT to determine
how accurate it's guess is. (Essentially mimicing
a real player reading the screen) However this
proved unviable. Thus the computer now references
the game object directly.